pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vwrapper"
version = "0.1.0"
description = "AI-powered VMware infrastructure control from the CLI."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
keywords = ["vmware", "vcenter", "cli", "ai", "infrastructure", "llm", "devops"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "typer>=0.12.3",
    "rich>=13.0.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "pyvmomi>=8.0.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-mock>=3.12.0"]

[project.scripts]
vwrapper = "vwrapper.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
